Some caching values for the prompt system. Set up framework for persistent inventory spells. Fix spawn eggs. Refactorings, cleanups. Added range option to nova effect, minimum value is 1, default value is 20. Note that the effective distance possible is limited by block render distance. Fixed AI flag on 1.11
Added a new option to drainlife spell called smoke which accepts a boolean, defaulting to true. When set to false, smoke will not play for the animation.
DrainLife spell now plays the special position of effects where it normally shows smoke. show-spell-effect must be true in order for this to work.

Finished moving offhand/mainhand details to HandHandler and out of the volatile code handles (cleanup is good, large volatile features files are annoying).

+++ src/com/nisovin/magicspells/spells/buff/FlamewalkSpell.java
@@ -94,7 +94,7 @@ public class FlamewalkSpell extends BuffSpell {
 						if (entity instanceof Player) {
 							if (entity != player && targetPlayers) {
 								if (checkPlugins) {
-									MagicSpellsEntityDamageByEntityEvent event = new MagicSpellsEntityDamageByEntityEvent(player, entity, DamageCause.ENTITY_ATTACK, (double)1);
+									MagicSpellsEntityDamageByEntityEvent event = new MagicSpellsEntityDamageByEntityEvent(player, entity, DamageCause.ENTITY_ATTACK, 1D);
 									Bukkit.getServer().getPluginManager().callEvent(event);
 									if (event.isCancelled()) {
 										continue;
